Move navigation file
