+++ lib/screens/app/ecosystem/ecosystem.dart
@@ -1,7 +1,7 @@
 
 
 import 'package:flutter/material.dart';
-import 'package:seeds/providers/services/navigation_service.dart';
+import 'package:seeds/v2/navigation/navigation_service.dart';
 
 class Ecosystem extends StatelessWidget {
   @override
