+++ lib/screens/app/ecosystem/proposals/proposals.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:seeds/models/models.dart';
 import 'package:seeds/providers/services/http_service.dart';
-import 'package:seeds/providers/services/navigation_service.dart';
+import 'package:seeds/v2/navigation/navigation_service.dart';
 import 'package:seeds/screens/app/ecosystem/proposals/proposal_header_details.dart';
 import 'package:seeds/i18n/proposals.i18n.dart';
 import 'package:seeds/utils/old_toolbox/paginated_list_view.dart';
