+++ lib/screens/app/guardians/guardian_invite_sent.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:seeds/i18n/guardians.i18n.dart';
-import 'package:seeds/providers/services/navigation_service.dart';
+import 'package:seeds/v2/navigation/navigation_service.dart';
 import 'package:seeds/widgets/main_button.dart';
 
 class InviteGuardiansSent extends StatelessWidget {
