+++ lib/screens/app/guardians/my_guardians_tutorial.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:seeds/providers/notifiers/settings_notifier.dart';
-import 'package:seeds/providers/services/navigation_service.dart';
+import 'package:seeds/v2/navigation/navigation_service.dart';
 import 'package:seeds/widgets/main_button.dart';
 
 Future<void> showFirstTimeUserDialog(BuildContext buildContext) async {
