+++ lib/screens/app/wallet/receive_qr.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart' hide Action;
 import 'package:seeds/providers/services/eos_service.dart';
-import 'package:seeds/providers/services/navigation_service.dart';
+import 'package:seeds/v2/navigation/navigation_service.dart';
 import 'package:seeds/widgets/main_button.dart';
 import '../../../utils/double_extension.dart';
 import 'package:qr_flutter/qr_flutter.dart';
