+++ lib/screens/app/wallet/transfer/transfer_form.dart
@@ -7,7 +7,7 @@ import 'package:provider/provider.dart';
 import 'package:seeds/v2/constants/app_colors.dart';
 import 'package:seeds/providers/notifiers/balance_notifier.dart';
 import 'package:seeds/providers/services/eos_service.dart';
-import 'package:seeds/providers/services/navigation_service.dart';
+import 'package:seeds/v2/navigation/navigation_service.dart';
 import 'package:seeds/screens/app/profile/image_viewer.dart';
 import 'package:seeds/v2/datasource/remote/model/balance_model.dart';
 import 'package:seeds/widgets/amount_field.dart';
