+++ lib/v2/screens/profile_screens/profile/components/profile_middle.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:seeds/v2/constants/app_colors.dart';
 import 'package:seeds/design/app_theme.dart';
-import 'package:seeds/providers/services/navigation_service.dart';
+import 'package:seeds/v2/navigation/navigation_service.dart';
 import 'package:seeds/v2/datasource/local/settings_storage.dart';
 import 'package:seeds/i18n/profile.i18n.dart';
 import 'package:seeds/v2/components/divider_jungle.dart';
