+++ lib/v2/screens/transfer/send_search_user/send_search_user_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:seeds/v2/components/search_user/search_user_widget.dart';
 import 'package:seeds/v2/datasource/remote/model/member_model.dart';
-import 'package:seeds/providers/services/navigation_service.dart';
+import 'package:seeds/v2/navigation/navigation_service.dart';
 
 /// SendSearchUserScreen SCREEN
 class SendSearchUserScreen extends StatelessWidget {
